fix(worker): fetch the host keys from admin.fedoraproject.org
As it appears, there is some ongoing issue with `ssh-keyscan` that blocks
running this action. The host keys are also present at a specific subpage
of admin.fedoraproject.org, so fetch them from there until this issue is
resolved.

Also pods trying to retry "bootstrap" can be in issue because of the
Fedora's DDoS protection.

I've also considered "baking in" the known hosts, but that could result
in issues during the image build.

Signed-off-by: Matej Focko <mfocko@packit.dev>