perf: share Copr client in periodic babysit run
Currently each call to ‹update_copr_builds› creates a new instance of
‹CoprClient› from the config file, this involves IO operations (lookup
and opening of the config), constructing the client (parsing the config
+ constructing proxy objects where each has its own ‹Requests› instance).

This can slow down the babysit task as a whole. Therefore create one
shared ‹CoprClient› instance that can be used for checking all
of the pending Copr builds.

Signed-off-by: Matej Focko <mfocko@packit.dev>