Trigger packit only for specific package in monorepo
Packit-service now supports the `--package` parameter
in PR comments to specify package for which to run jobs. One test
was added to test this functionality.

The argparse module is now used for parsing comments.
By default, argparse seems to ignore newlines in the epilog string
used when generating help message, resulting in a messy help
message. RawTextHelpFormatter ensures the message (specifically,
the epilog part) is formatted properly. This going to be needed
when argparse is utilized by `/packit help` command in the future.

A collection of simple tests were also added to test the comment
parsers.