Filter out Copr builds without SRPM in SQL
In addition to pagination slicing, the CoprBuildsList class uses Python
code to filter out Copr builds that are waiting for an SRPM or whose
SRPM build failed. This causes that the API in some cases can return
fewer items than the user requested.

Moving the filter into SQL resolves the problem.

Resolves: https://github.com/packit/packit-service/issues/2505